refactor(transit): introduce TransitPlanner and move the MOTIS planner behind it
