+++ server/src/nest/transit/transit.module.ts
@@ -3,6 +3,7 @@ import { Module } from '@nestjs/common';
 import { TransitController } from './transit.controller';
 import { TransitService } from './transit.service';
 import { TransitMcp } from './transit.mcp';
+import { TransitousPlanner } from './providers/transitous.planner';
 import { DaysModule } from '../days/days.module';
 import { ReservationsModule } from '../reservations/reservations.module';
 import { AuthModule } from '../auth/auth.module';
@@ -17,7 +18,7 @@ import { McpSharedModule } from '../mcp-shared/mcp-shared.module';
   // DaysModule + ReservationsModule: TransitMcp's create_transit_journey injects both.
   imports: [McpSharedModule, RateLimitModule, DaysModule, ReservationsModule, AuthModule],
   controllers: [TransitController],
-  providers: [TransitService, TransitMcp],
+  providers: [TransitService, TransitMcp, TransitousPlanner],
   exports: [TransitService],
 })
 export class TransitModule {}
