+++ server/src/nest/transit/transit.service.ts
@@ -1,45 +1,22 @@
 import { Injectable } from '@nestjs/common';
-import { getAppUrl, readEnv } from '../../app-config';
-import { buildUserAgent } from '../maps/maps.helpers';
-import { badRequest, fetchJson } from './transit.http';
-import {
-  deriveTransitStats,
-  POLYLINE_PRECISION,
-  safeColor,
-  SCHEDULED_TRANSIT_MODES,
-  type PlanQuery,
-  type TransitItinerary,
-  type TransitLeg,
-  type TransitLegStop,
-  type TransitPlace,
-} from './transit.helpers';
+import { badRequest } from './transit.http';
+import { SCHEDULED_TRANSIT_MODES, type PlanQuery, type TransitPlace } from './transit.helpers';
+import type { TransitPlanResult, TransitPlanner, TransitProvider } from './providers/transit-planner';
+import { TransitousPlanner, upstream } from './providers/transitous.planner';
 
 /**
- * Public transit routing (#1065) backed by Transitous (api.transitous.org), the
- * community-run MOTIS instance over public GTFS feeds — free, no API key, fits
- * TREK's no-paid-providers rule. Self-hosters can point TRANSIT_API_URL at their
- * own MOTIS instance instead.
+ * Public transit routing (#1065). This service owns what is provider-agnostic —
+ * query validation, the shared response cache, and the dispatch to the planner
+ * the administrator selected — while each provider's request building and
+ * response mapping lives in providers/.
  *
- * This service is a thin, validating proxy: the browser never talks to
- * Transitous directly (their usage policy wants an identifying User-Agent with
- * contact info, which we send once from the server), and responses are mapped
- * to a compact shape so the client isn't coupled to the MOTIS schema.
+ * `geocode` is NOT dispatched: it always goes to Transitous, because the
+ * NAVITIME subscription exposes no geocoding endpoint at all.
  *
- * Folded 1:1 from the legacy services/transitService.ts. The base URL, the
- * lazy User-Agent memo and the response cache stay module-scoped on purpose
- * (the permissions/exchange-rates precedent): the DI singleton and any future
- * non-Nest caller share one cache, and the base URL keeps its frozen-at-import
- * legacy timing.
+ * The response cache stays module-scoped on purpose (the
+ * permissions/exchange-rates precedent): the DI singleton and any future
+ * non-Nest caller share one cache.
  */
-
-// Frozen at import on purpose (legacy timing).
-const TRANSIT_API_BASE = readEnv().integrations.transitApiBase;
-let userAgent: string | null = null;
-
-function getUserAgent(): string {
-  userAgent ??= buildUserAgent(getAppUrl());
-  return userAgent;
-}
 
 const ALLOWED_MODES = new Set(['TRANSIT', ...SCHEDULED_TRANSIT_MODES]);
 
@@ -79,43 +56,14 @@ function isCoord(v: string): boolean {
   return Math.abs(lat) <= 90 && Math.abs(lng) <= 180;
 }
 
-/** Transitous keeps its base URL and its identifying User-Agent; the rest of the
- * upstream contract (timeout, size ceiling, 429/502) lives in transit.http.ts. */
-async function upstream(path: string, params: URLSearchParams): Promise<unknown> {
-  return fetchJson(`${TRANSIT_API_BASE}${path}?${params}`, { 'User-Agent': getUserAgent() });
-}
-
-interface MotisPlaceRaw {
-  name?: string;
-  lat?: number;
-  lon?: number;
-  departure?: string;
-  arrival?: string;
-  scheduledDeparture?: string;
-  scheduledArrival?: string;
-  track?: string;
-  scheduledTrack?: string;
-}
-
-function mapStop(p: MotisPlaceRaw | undefined, kind: 'departure' | 'arrival'): TransitLegStop {
-  return {
-    name: p?.name || '',
-    // A missing coordinate becomes 0 (null island) rather than null on
-    // purpose: TransitLegStop.lat/lng are non-nullable across the wire
-    // contract (transitStopSchema, the client picker) — making them nullable
-    // would ripple through transit-itinerary.helpers' validation and drop
-    // otherwise-usable itineraries. MOTIS places carry coordinates in
-    // practice; endpoint stops are re-checked by transitCoordinatesMatch.
-    lat: typeof p?.lat === 'number' ? p.lat : 0,
-    lng: typeof p?.lon === 'number' ? p.lon : 0,
-    time: (kind === 'departure' ? p?.departure : p?.arrival) || null,
-    scheduledTime: (kind === 'departure' ? p?.scheduledDeparture : p?.scheduledArrival) || null,
-    track: p?.track || p?.scheduledTrack || null,
-  };
-}
-
 @Injectable()
 export class TransitService {
+  private readonly planners: { transitous: TransitPlanner };
+
+  constructor(transitous: TransitousPlanner) {
+    this.planners = { transitous };
+  }
+
   /** Station/place search for the from/to pickers. `near` biases results. */
   async geocode(query: string, language?: string, near?: string): Promise<{ results: TransitPlace[] }> {
     const text = (query || '').trim();
@@ -149,103 +97,40 @@ export class TransitService {
     return data;
   }
 
-  /** Route search between two coordinates. Returns compact itineraries for the picker. */
-  async plan(q: PlanQuery): Promise<{ itineraries: TransitItinerary[] }> {
-    const bad = (msg: string): never => {
-      throw badRequest(msg);
-    };
-    if (!q.from || !isCoord(q.from)) bad('from must be "lat,lng"');
-    if (!q.to || !isCoord(q.to)) bad('to must be "lat,lng"');
-
-    const params = new URLSearchParams({ fromPlace: q.from, toPlace: q.to, numItineraries: '8' });
-
-    if (q.time) {
-      const parsed = new Date(q.time);
-      if (Number.isNaN(parsed.getTime())) bad('time must be an ISO date-time');
-      params.set('time', parsed.toISOString());
+  /**
+   * Route search between two coordinates. Validates, then hands the query to the
+   * selected provider — every 400 below is raised here so no provider has to
+   * repeat it.
+   */
+  async plan(q: PlanQuery): Promise<TransitPlanResult> {
+    if (!q.from || !isCoord(q.from)) throw badRequest('from must be "lat,lng"');
+    if (!q.to || !isCoord(q.to)) throw badRequest('to must be "lat,lng"');
+    if (q.time !== undefined && Number.isNaN(new Date(q.time).getTime())) {
+      throw badRequest('time must be an ISO date-time');
     }
-    if (q.arriveBy) params.set('arriveBy', 'true');
-
     if (q.modes) {
       const modes = q.modes
         .split(',')
         .map((m) => m.trim().toUpperCase())
         .filter(Boolean);
-      if (modes.some((m) => !ALLOWED_MODES.has(m))) bad('unsupported transit mode');
-      if (modes.length > 0) params.set('transitModes', modes.join(','));
+      if (modes.some((m) => !ALLOWED_MODES.has(m))) throw badRequest('unsupported transit mode');
     }
     if (q.maxTransfers !== undefined && q.maxTransfers !== null) {
       const n = Number(q.maxTransfers);
-      if (!Number.isInteger(n) || n < 0 || n > 10) bad('maxTransfers must be 0-10');
-      params.set('maxTransfers', String(n));
+      if (!Number.isInteger(n) || n < 0 || n > 10) throw badRequest('maxTransfers must be 0-10');
     }
-    // We only want scheduled transit journeys in the results — a pure-walk
-    // "direct" connection is what the existing OSRM footpath routing is for.
-    params.set('directModes', 'WALK');
 
-    const key = `plan:${params.toString()}`;
+    const provider: TransitProvider = 'transitous';
+    const planner = this.planners[provider];
+
+    // Keyed on the TREK query plus the provider id, not on the provider's own
+    // parameters: without the id, flipping providers would serve the previous
+    // one's itineraries for a full TTL.
+    const key = `plan:${planner.id}:${q.from}|${q.to}|${q.time ?? ''}|${q.arriveBy ? '1' : '0'}|${q.modes ?? ''}|${q.maxTransfers ?? ''}`;
     const cached = cacheGet(key);
-    if (cached) return cached as { itineraries: TransitItinerary[] };
+    if (cached) return cached as TransitPlanResult;
 
-    const raw = (await upstream('/api/v6/plan', params)) as {
-      itineraries?: Array<{
-        duration?: number;
-        startTime?: string;
-        endTime?: string;
-        transfers?: number;
-        legs?: Array<{
-          mode?: string;
-          duration?: number;
-          distance?: number;
-          headsign?: string;
-          routeShortName?: string;
-          displayName?: string;
-          routeColor?: string;
-          routeTextColor?: string;
-          agencyName?: string;
-          from?: MotisPlaceRaw;
-          to?: MotisPlaceRaw;
-          intermediateStops?: unknown[];
-          legGeometry?: { points?: string; precision?: number };
-        }>;
-      }>;
-    };
-
-    const itineraries: TransitItinerary[] = (raw.itineraries || []).flatMap((it) => {
-      if (!it.startTime || !it.endTime || !Array.isArray(it.legs)) return [];
-      const legs: TransitLeg[] = it.legs.map((leg) => ({
-        mode: (leg.mode || 'WALK').toUpperCase(),
-        from: mapStop(leg.from, 'departure'),
-        to: mapStop(leg.to, 'arrival'),
-        duration: typeof leg.duration === 'number' ? leg.duration : 0,
-        distance: typeof leg.distance === 'number' ? Math.round(leg.distance) : null,
-        headsign: leg.headsign || null,
-        // displayName is the public identifier MOTIS resolved for the run, taking
-        // the trip number into account where the feed has one. Plenty of operators
-        // (German long distance among them) keep routeShortName as an internal line
-        // number, so preferring it showed "20" instead of "ICE 72" (#1715).
-        line: leg.displayName || leg.routeShortName || null,
-        lineColor: safeColor(leg.routeColor),
-        lineTextColor: safeColor(leg.routeTextColor),
-        agency: leg.agencyName || null,
-        intermediateStops: Array.isArray(leg.intermediateStops) ? leg.intermediateStops.length : 0,
-        geometry: leg.legGeometry?.points || null,
-        geometryPrecision: leg.legGeometry?.precision ?? POLYLINE_PRECISION,
-      }));
-      const stats = deriveTransitStats(it.startTime, it.endTime, legs, it.transfers);
-      return [
-        {
-          startTime: it.startTime,
-          endTime: it.endTime,
-          // Wall-clock duration (start→end) so waits/transfers count — summing leg
-          // run-times would understate the journey and mis-slot it in the timeline.
-          ...stats,
-          legs,
-        },
-      ];
-    });
-
-    const data = { itineraries };
+    const data = await planner.plan(q);
     cacheSet(key, data);
     return data;
   }
